Update JDBC-using readers to more appropriately use JDBC
